@@ -17,7 +17,7 @@ limitations under the License.
 package options
 
 import (
-	"flag"
+	flag "github.com/spf13/pflag"
 
 	cliflag "k8s.io/component-base/cli/flag"
 )
@@ -37,6 +37,10 @@ type ControllerOptions struct {
 	AwsSdkDebugLog bool
 	// flag to warn on invalid tag, instead of returning an error
 	WarnOnInvalidTag bool
+	// flag to set user agent
+	UserAgentExtra string
+	// flag to enable batching of API calls
+	Batching bool
 }
 
 func (s *ControllerOptions) AddFlags(fs *flag.FlagSet) {
@@ -45,4 +49,6 @@ func (s *ControllerOptions) AddFlags(fs *flag.FlagSet) {
 	fs.StringVar(&s.KubernetesClusterID, "k8s-tag-cluster-id", "", "ID of the Kubernetes cluster used for tagging provisioned EBS volumes (optional).")
 	fs.BoolVar(&s.AwsSdkDebugLog, "aws-sdk-debug-log", false, "To enable the aws sdk debug log level (default to false).")
 	fs.BoolVar(&s.WarnOnInvalidTag, "warn-on-invalid-tag", false, "To warn on invalid tags, instead of returning an error")
+	fs.StringVar(&s.UserAgentExtra, "user-agent-extra", "", "Extra string appended to user agent.")
+	fs.BoolVar(&s.Batching, "batching", false, "To enable batching of API calls. This is especially helpful for improving performance in workloads that are sensitive to EC2 rate limits.")
 }
