@@ -17,11 +17,12 @@ limitations under the License.
 package main
 
 import (
-	"flag"
 	"os"
 	"reflect"
 	"strconv"
 	"testing"
+
+	flag "github.com/spf13/pflag"
 
 	"github.com/kubernetes-sigs/aws-ebs-csi-driver/pkg/driver"
 )
@@ -50,26 +51,34 @@ func TestGetOptions(t *testing.T) {
 		awsSdkDebugFlagValue := true
 		VolumeAttachLimitFlagName := "volume-attach-limit"
 		var VolumeAttachLimit int64 = 42
+		userAgentExtraFlag := "user-agent-extra"
+		userAgentExtraFlagValue := "test"
+		otelTracingFlagName := "enable-otel-tracing"
+		otelTracingFlagValue := true
+		batchingFlagName := "batching"
+		batchingFlagValue := true
 
 		args := append([]string{
 			"aws-ebs-csi-driver",
 		}, additionalArgs...)
 
 		if withServerOptions {
-			args = append(args, "-"+endpointFlagName+"="+endpoint)
+			args = append(args, "--"+endpointFlagName+"="+endpoint)
+			args = append(args, "--"+otelTracingFlagName+"="+strconv.FormatBool(otelTracingFlagValue))
 		}
 		if withControllerOptions {
-			args = append(args, "-"+extraTagsFlagName+"="+extraTagKey+"="+extraTagValue)
-			args = append(args, "-"+awsSdkDebugFlagName+"="+strconv.FormatBool(awsSdkDebugFlagValue))
+			args = append(args, "--"+extraTagsFlagName+"="+extraTagKey+"="+extraTagValue)
+			args = append(args, "--"+awsSdkDebugFlagName+"="+strconv.FormatBool(awsSdkDebugFlagValue))
+			args = append(args, "--"+userAgentExtraFlag+"="+userAgentExtraFlagValue)
+			args = append(args, "--"+batchingFlagName+"="+strconv.FormatBool(batchingFlagValue))
 		}
 		if withNodeOptions {
-			args = append(args, "-"+VolumeAttachLimitFlagName+"="+strconv.FormatInt(VolumeAttachLimit, 10))
+			args = append(args, "--"+VolumeAttachLimitFlagName+"="+strconv.FormatInt(VolumeAttachLimit, 10))
 		}
 
 		oldArgs := os.Args
 		defer func() { os.Args = oldArgs }()
 		os.Args = args
-
 		options := GetOptions(flagSet)
 
 		if withServerOptions {
@@ -79,6 +88,10 @@ func TestGetOptions(t *testing.T) {
 			}
 			if options.ServerOptions.Endpoint != endpoint {
 				t.Fatalf("expected endpoint to be %q but it is %q", endpoint, options.ServerOptions.Endpoint)
+			}
+			otelTracingFlag := flagSet.Lookup(otelTracingFlagName)
+			if otelTracingFlag == nil {
+				t.Fatalf("expected %q flag to be added but it is not", otelTracingFlagName)
 			}
 		}
 
@@ -96,6 +109,16 @@ func TestGetOptions(t *testing.T) {
 			}
 			if options.ControllerOptions.AwsSdkDebugLog != awsSdkDebugFlagValue {
 				t.Fatalf("expected sdk debug flag to be %v but it is %v", awsSdkDebugFlagValue, options.ControllerOptions.AwsSdkDebugLog)
+			}
+			if options.ControllerOptions.UserAgentExtra != userAgentExtraFlagValue {
+				t.Fatalf("expected user agent string to be %q but it is %q", userAgentExtraFlagValue, options.ControllerOptions.UserAgentExtra)
+			}
+			batchingFlag := flagSet.Lookup(batchingFlagName)
+			if batchingFlag == nil {
+				t.Fatalf("expected %q flag to be added but it is not", batchingFlagName)
+			}
+			if options.ControllerOptions.Batching != batchingFlagValue {
+				t.Fatalf("expected sdk debug flag to be %v but it is %v", batchingFlagValue, options.ControllerOptions.Batching)
 			}
 		}
 
@@ -163,8 +186,10 @@ func TestGetOptions(t *testing.T) {
 				defer func() { osExit = oldOSExit }()
 
 				var exitCode int
+				calledExit := false
 				testExit := func(code int) {
 					exitCode = code
+					calledExit = true
 				}
 				osExit = testExit
 
@@ -172,7 +197,7 @@ func TestGetOptions(t *testing.T) {
 				defer func() { os.Args = oldArgs }()
 				os.Args = []string{
 					"aws-ebs-csi-driver",
-					"-version",
+					"--version",
 				}
 
 				flagSet := flag.NewFlagSet("test-flagset", flag.ContinueOnError)
@@ -180,6 +205,9 @@ func TestGetOptions(t *testing.T) {
 
 				if exitCode != 0 {
 					t.Fatalf("expected exit code 0 but got %d", exitCode)
+				}
+				if !calledExit {
+					t.Fatalf("expect osExit to be called, but wasn't")
 				}
 			},
 		},
