@@ -22,8 +22,7 @@ import (
 
 // ExistingNames is a map of assigned device names. Presence of a key with a device
 // name in the map means that the device is allocated. Value is irrelevant and
-// can be used for anything that NameAllocator user wants.  Only the relevant
-// part of device name should be in the map, e.g. "ba" for "/dev/xvdba".
+// can be used for anything that NameAllocator user wants.
 type ExistingNames map[string]string
 
 // On AWS, we should assign new (not yet used) device names to attached volumes.
@@ -35,9 +34,6 @@ type ExistingNames map[string]string
 // call), so all available device names are used eventually and it minimizes
 // device name reuse.
 type NameAllocator interface {
-	// GetNext returns a free device name or error when there is no free device
-	// name. Only the device name is returned, e.g. "ba" for "/dev/xvdba".
-	// It's up to the called to add appropriate "/dev/sd" or "/dev/xvd" prefix.
 	GetNext(existingNames ExistingNames) (name string, err error)
 }
 
@@ -45,19 +41,12 @@ type nameAllocator struct{}
 
 var _ NameAllocator = &nameAllocator{}
 
-// GetNext gets next available device given existing names that are being used
-// This function iterate through the device names in deterministic order of:
-//
-//	ba, ... ,bz, ca, ... , cz
-//
-// and return the first one that is not used yet.
+// GetNext returns a free device name or error when there is no free device name
+// It does this by using a list of legal EBS device names from device_names.go
 func (d *nameAllocator) GetNext(existingNames ExistingNames) (string, error) {
-	for _, c1 := range []string{"b", "c"} {
-		for c2 := 'a'; c2 <= 'z'; c2++ {
-			name := fmt.Sprintf("%s%s", c1, string(c2))
-			if _, found := existingNames[name]; !found {
-				return name, nil
-			}
+	for _, name := range deviceNames {
+		if _, found := existingNames[name]; !found {
+			return name, nil
 		}
 	}
 
