@@ -22,9 +22,9 @@ var DefaultEC2MetadataClient = func() (EC2Metadata, error) {
 
 func EC2MetadataInstanceInfo(svc EC2Metadata, regionFromSession string) (*Metadata, error) {
 	doc, err := svc.GetInstanceIdentityDocument()
-	klog.Infof("regionFromSession %v", regionFromSession)
+	klog.InfoS("Retrieving EC2 instance identity metadata", "regionFromSession", regionFromSession)
 	if err != nil {
-		return nil, fmt.Errorf("could not get EC2 instance identity metadata: %v", err)
+		return nil, fmt.Errorf("could not get EC2 instance identity metadata: %w", err)
 	}
 
 	if len(doc.InstanceID) == 0 {
@@ -53,7 +53,7 @@ func EC2MetadataInstanceInfo(svc EC2Metadata, regionFromSession string) (*Metada
 
 	enis, err := svc.GetMetadata(enisEndpoint)
 	if err != nil {
-		return nil, fmt.Errorf("could not get number of attached ENIs: %v", err)
+		return nil, fmt.Errorf("could not get number of attached ENIs: %w", err)
 	}
 	// the ENIs should not be empty; if (somehow) it is empty, return an error
 	if enis == "" {
@@ -61,17 +61,14 @@ func EC2MetadataInstanceInfo(svc EC2Metadata, regionFromSession string) (*Metada
 	}
 
 	attachedENIs := strings.Count(enis, "\n") + 1
-
-	//As block device mapping contains 1 volume for the AMI.
-	blockDevMappings := 1
+	blockDevMappings := 0
 
 	if !util.IsSBE(doc.Region) {
-		mappings, err := svc.GetMetadata(blockDevicesEndpoint)
-		// The output contains 1 volume for the AMI. Any other block device contributes to the attachment limit
-		blockDevMappings = strings.Count(mappings, "\n")
-		if err != nil {
-			return nil, fmt.Errorf("could not get number of block device mappings: %v", err)
+		mappings, mapErr := svc.GetMetadata(blockDevicesEndpoint)
+		if mapErr != nil {
+			return nil, fmt.Errorf("could not get number of block device mappings: %w", err)
 		}
+		blockDevMappings = strings.Count(mappings, "ebs")
 	}
 
 	instanceInfo := Metadata{
@@ -88,15 +85,15 @@ func EC2MetadataInstanceInfo(svc EC2Metadata, regionFromSession string) (*Metada
 	// it's guaranteed to be in the form `arn:<partition>:outposts:<region>:<account>:outpost/<outpost-id>`
 	// There's a case to be made here to ignore the error so a failure here wouldn't affect non-outpost calls.
 	if err != nil && !strings.Contains(err.Error(), "404") {
-		return nil, fmt.Errorf("something went wrong while getting EC2 outpost arn: %s", err.Error())
+		return nil, fmt.Errorf("something went wrong while getting EC2 outpost arn: %w", err)
 	} else if err == nil {
-		klog.Infof("Running in an outpost environment with arn: %s", outpostArn)
+		klog.InfoS("Running in an outpost environment with arn", "outpostArn", outpostArn)
 		outpostArn = strings.ReplaceAll(outpostArn, "outpost/", "")
 		parsedArn, err := arn.Parse(outpostArn)
 		if err != nil {
-			klog.Warningf("Failed to parse the outpost arn: %s", outpostArn)
+			klog.InfoS("Failed to parse the outpost arn", "outpostArn", outpostArn)
 		} else {
-			klog.Infof("Using outpost arn: %v", parsedArn)
+			klog.InfoS("Using outpost arn", "parsedArn", parsedArn)
 			instanceInfo.OutpostArn = parsedArn
 		}
 	}
