@@ -18,33 +18,32 @@ package driver
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"os"
 	"strconv"
 	"strings"
 
 	"github.com/aws/aws-sdk-go/aws/arn"
+	"github.com/awslabs/volume-modifier-for-k8s/pkg/rpc"
 	csi "github.com/container-storage-interface/spec/lib/go/csi"
-	"github.com/golang/protobuf/ptypes"
 	"github.com/kubernetes-sigs/aws-ebs-csi-driver/pkg/cloud"
 	"github.com/kubernetes-sigs/aws-ebs-csi-driver/pkg/driver/internal"
 	"github.com/kubernetes-sigs/aws-ebs-csi-driver/pkg/util"
 	"github.com/kubernetes-sigs/aws-ebs-csi-driver/pkg/util/template"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+	"google.golang.org/protobuf/types/known/timestamppb"
 	"k8s.io/klog/v2"
 )
 
-var (
-	// volumeCaps represents how the volume could be accessed.
-	// It is SINGLE_NODE_WRITER since EBS volume could only be
-	// attached to a single node at any given time.
-	volumeCaps = []csi.VolumeCapability_AccessMode{
-		{
-			Mode: csi.VolumeCapability_AccessMode_SINGLE_NODE_WRITER,
-		},
-	}
+// Supported access modes
+const (
+	SingleNodeWriter     = csi.VolumeCapability_AccessMode_SINGLE_NODE_WRITER
+	MultiNodeMultiWriter = csi.VolumeCapability_AccessMode_MULTI_NODE_MULTI_WRITER
+)
 
+var (
 	// controllerCaps represents the capability of controller service
 	controllerCaps = []csi.ControllerServiceCapability_RPC_Type{
 		csi.ControllerServiceCapability_RPC_CREATE_DELETE_VOLUME,
@@ -59,9 +58,12 @@ const isManagedByDriver = "true"
 
 // controllerService represents the controller service of CSI driver
 type controllerService struct {
-	cloud         cloud.Cloud
-	inFlight      *internal.InFlight
-	driverOptions *DriverOptions
+	cloud               cloud.Cloud
+	inFlight            *internal.InFlight
+	driverOptions       *DriverOptions
+	modifyVolumeManager *modifyVolumeManager
+
+	rpc.UnimplementedModifyServer
 }
 
 var (
@@ -78,29 +80,31 @@ var (
 func newControllerService(driverOptions *DriverOptions) controllerService {
 	region := os.Getenv("AWS_REGION")
 	if region == "" {
-		klog.V(5).Infof("[Debug] Retrieving region from metadata service")
+		klog.V(5).InfoS("[Debug] Retrieving region from metadata service")
 		metadata, err := NewMetadataFunc(cloud.DefaultEC2MetadataClient, cloud.DefaultKubernetesAPIClient, region)
 		if err != nil {
-			klog.Errorf("Could not determine region from any metadata service. The region can be manually supplied via the AWS_REGION environment variable.")
+			klog.ErrorS(err, "Could not determine region from any metadata service. The region can be manually supplied via the AWS_REGION environment variable.")
 			panic(err)
 		}
 		region = metadata.GetRegion()
 	}
 
-	cloudSrv, err := NewCloudFunc(region, driverOptions.awsSdkDebugLog)
+	klog.InfoS("batching", "status", driverOptions.batching)
+	cloudSrv, err := NewCloudFunc(region, driverOptions.awsSdkDebugLog, driverOptions.userAgentExtra, driverOptions.batching)
 	if err != nil {
 		panic(err)
 	}
 
 	return controllerService{
-		cloud:         cloudSrv,
-		inFlight:      internal.NewInFlight(),
-		driverOptions: driverOptions,
+		cloud:               cloudSrv,
+		inFlight:            internal.NewInFlight(),
+		driverOptions:       driverOptions,
+		modifyVolumeManager: newModifyVolumeManager(),
 	}
 }
 
 func (d *controllerService) CreateVolume(ctx context.Context, req *csi.CreateVolumeRequest) (*csi.CreateVolumeResponse, error) {
-	klog.V(4).Infof("CreateVolume: called with args %+v", *req)
+	klog.V(4).InfoS("CreateVolume: called", "args", *req)
 	if err := validateCreateVolumeRequest(req); err != nil {
 		return nil, err
 	}
@@ -109,6 +113,15 @@ func (d *controllerService) CreateVolume(ctx context.Context, req *csi.CreateVol
 		return nil, err
 	}
 	volName := req.GetName()
+	volCap := req.GetVolumeCapabilities()
+
+	multiAttach := false
+	for _, c := range volCap {
+		if c.GetAccessMode().GetMode() == MultiNodeMultiWriter && isBlock(c) {
+			klog.V(4).InfoS("CreateVolume: multi-attach is enabled", "volumeID", volName)
+			multiAttach = true
+		}
+	}
 
 	// check if a request is already in-flight
 	if ok := d.inFlight.Insert(volName); !ok {
@@ -118,28 +131,33 @@ func (d *controllerService) CreateVolume(ctx context.Context, req *csi.CreateVol
 	defer d.inFlight.Delete(volName)
 
 	var (
-		volumeType              string
-		iopsPerGB               int
-		allowIOPSPerGBIncrease  bool
-		reconcileGP3Performance bool
-		iops                    int
-		throughput              int
-		isEncrypted             bool
-		blockExpress            bool
-		kmsKeyID                string
-		scTags                  []string
-		volumeTags              = map[string]string{
+		volumeType             string
+		iopsPerGB              int
+		allowIOPSPerGBIncrease bool
+		iops                   int
+		throughput             int
+		isEncrypted            bool
+		blockExpress           bool
+		kmsKeyID               string
+		scTags                 []string
+		volumeTags             = map[string]string{
 			cloud.VolumeNameTagKey:   volName,
 			cloud.AwsEbsDriverTagKey: isManagedByDriver,
 		}
+		blockSize       string
+		inodeSize       string
+		bytesPerInode   string
+		numberOfInodes  string
+		ext4BigAlloc    bool
+		ext4ClusterSize string
 	)
 
-	tProps := new(template.Props)
+	tProps := new(template.PVProps)
 
 	for key, value := range req.GetParameters() {
 		switch strings.ToLower(key) {
 		case "fstype":
-			klog.Warning("\"fstype\" is deprecated, please use \"csi.storage.k8s.io/fstype\" instead")
+			klog.InfoS("\"fstype\" is deprecated, please use \"csi.storage.k8s.io/fstype\" instead")
 		case VolumeTypeKey:
 			volumeType = value
 		case IopsPerGBKey:
@@ -149,8 +167,6 @@ func (d *controllerService) CreateVolume(ctx context.Context, req *csi.CreateVol
 			}
 		case AllowAutoIOPSPerGBIncreaseKey:
 			allowIOPSPerGBIncrease = value == "true"
-		case ReconcileGP3PerformanceKey:
-			reconcileGP3Performance = value == "true"
 		case IopsKey:
 			iops, err = strconv.Atoi(value)
 			if err != nil {
@@ -180,6 +196,35 @@ func (d *controllerService) CreateVolume(ctx context.Context, req *csi.CreateVol
 			if value == "true" {
 				blockExpress = true
 			}
+		case BlockSizeKey:
+			if isAlphanumeric := util.StringIsAlphanumeric(value); !isAlphanumeric {
+				return nil, status.Errorf(codes.InvalidArgument, "Could not parse blockSize (%s): %v", value, err)
+			}
+			blockSize = value
+		case InodeSizeKey:
+			if isAlphanumeric := util.StringIsAlphanumeric(value); !isAlphanumeric {
+				return nil, status.Errorf(codes.InvalidArgument, "Could not parse inodeSize (%s): %v", value, err)
+			}
+			inodeSize = value
+		case BytesPerInodeKey:
+			if isAlphanumeric := util.StringIsAlphanumeric(value); !isAlphanumeric {
+				return nil, status.Errorf(codes.InvalidArgument, "Could not parse bytesPerInode (%s): %v", value, err)
+			}
+			bytesPerInode = value
+		case NumberOfInodesKey:
+			if isAlphanumeric := util.StringIsAlphanumeric(value); !isAlphanumeric {
+				return nil, status.Errorf(codes.InvalidArgument, "Could not parse numberOfInodes (%s): %v", value, err)
+			}
+			numberOfInodes = value
+		case Ext4BigAllocKey:
+			if value == "true" {
+				ext4BigAlloc = true
+			}
+		case Ext4ClusterSizeKey:
+			if isAlphanumeric := util.StringIsAlphanumeric(value); !isAlphanumeric {
+				return nil, status.Errorf(codes.InvalidArgument, "Could not parse ext4ClusterSize (%s): %v", value, err)
+			}
+			ext4ClusterSize = value
 		default:
 			if strings.HasPrefix(key, TagKeyPrefix) {
 				scTags = append(scTags, value)
@@ -189,10 +234,47 @@ func (d *controllerService) CreateVolume(ctx context.Context, req *csi.CreateVol
 		}
 	}
 
-	if volumeType == cloud.VolumeTypeIO1 {
-		if iopsPerGB == 0 {
-			return nil, status.Errorf(codes.InvalidArgument, "The parameter IOPSPerGB must be specified for io1 volumes")
+	responseCtx := map[string]string{}
+
+	if len(blockSize) > 0 {
+		responseCtx[BlockSizeKey] = blockSize
+		if err = validateFormattingOption(volCap, BlockSizeKey, FileSystemConfigs); err != nil {
+			return nil, err
 		}
+	}
+	if len(inodeSize) > 0 {
+		responseCtx[InodeSizeKey] = inodeSize
+		if err = validateFormattingOption(volCap, InodeSizeKey, FileSystemConfigs); err != nil {
+			return nil, err
+		}
+	}
+	if len(bytesPerInode) > 0 {
+		responseCtx[BytesPerInodeKey] = bytesPerInode
+		if err = validateFormattingOption(volCap, BytesPerInodeKey, FileSystemConfigs); err != nil {
+			return nil, err
+		}
+	}
+	if len(numberOfInodes) > 0 {
+		responseCtx[NumberOfInodesKey] = numberOfInodes
+		if err = validateFormattingOption(volCap, NumberOfInodesKey, FileSystemConfigs); err != nil {
+			return nil, err
+		}
+	}
+	if ext4BigAlloc {
+		responseCtx[Ext4BigAllocKey] = "true"
+		if err = validateFormattingOption(volCap, Ext4BigAllocKey, FileSystemConfigs); err != nil {
+			return nil, err
+		}
+	}
+	if len(ext4ClusterSize) > 0 {
+		responseCtx[Ext4ClusterSizeKey] = ext4ClusterSize
+		if err = validateFormattingOption(volCap, Ext4ClusterSizeKey, FileSystemConfigs); err != nil {
+			return nil, err
+		}
+	}
+
+	if !ext4BigAlloc && len(ext4ClusterSize) > 0 {
+		return nil, status.Errorf(codes.InvalidArgument, "Cannot set ext4BigAllocClusterSize when ext4BigAlloc is false")
 	}
 
 	if blockExpress && volumeType != cloud.VolumeTypeIO2 {
@@ -227,17 +309,12 @@ func (d *controllerService) CreateVolume(ctx context.Context, req *csi.CreateVol
 		volumeTags[k] = v
 	}
 
-	// add the reconciliation tag identifier to the volume
-	if volumeType == cloud.VolumeTypeGP3 && reconcileGP3Performance {
-		volumeTags[cloud.AwsEbsReconcileGP3PerformanceTagKey] = "true"
-	}
-
 	addTags, err := template.Evaluate(scTags, tProps, d.driverOptions.warnOnInvalidTag)
 	if err != nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Error interpolating the tag value: %v", err)
 	}
 
-	if err := validateExtraTags(addTags, d.driverOptions.warnOnInvalidTag); err != nil {
+	if err = validateExtraTags(addTags, d.driverOptions.warnOnInvalidTag); err != nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Invalid tag value: %v", err)
 	}
 
@@ -246,34 +323,34 @@ func (d *controllerService) CreateVolume(ctx context.Context, req *csi.CreateVol
 	}
 
 	opts := &cloud.DiskOptions{
-		CapacityBytes:           volSizeBytes,
-		Tags:                    volumeTags,
-		VolumeType:              volumeType,
-		IOPSPerGB:               iopsPerGB,
-		AllowIOPSPerGBIncrease:  allowIOPSPerGBIncrease,
-		ReconcileGP3Performance: reconcileGP3Performance,
-		IOPS:                    iops,
-		Throughput:              throughput,
-		AvailabilityZone:        zone,
-		OutpostArn:              outpostArn,
-		Encrypted:               isEncrypted,
-		BlockExpress:            blockExpress,
-		KmsKeyID:                kmsKeyID,
-		SnapshotID:              snapshotID,
+		CapacityBytes:          volSizeBytes,
+		Tags:                   volumeTags,
+		VolumeType:             volumeType,
+		IOPSPerGB:              iopsPerGB,
+		AllowIOPSPerGBIncrease: allowIOPSPerGBIncrease,
+		IOPS:                   iops,
+		Throughput:             throughput,
+		AvailabilityZone:       zone,
+		OutpostArn:             outpostArn,
+		Encrypted:              isEncrypted,
+		BlockExpress:           blockExpress,
+		KmsKeyID:               kmsKeyID,
+		SnapshotID:             snapshotID,
+		MultiAttachEnabled:     multiAttach,
 	}
 
 	disk, err := d.cloud.CreateDisk(ctx, volName, opts)
 	if err != nil {
 		errCode := codes.Internal
-		if err == cloud.ErrNotFound {
+		if errors.Is(err, cloud.ErrNotFound) {
 			errCode = codes.NotFound
 		}
-		if err == cloud.ErrIdempotentParameterMismatch {
+		if errors.Is(err, cloud.ErrIdempotentParameterMismatch) {
 			errCode = codes.AlreadyExists
 		}
 		return nil, status.Errorf(errCode, "Could not create volume %q: %v", volName, err)
 	}
-	return newCreateVolumeResponse(disk), nil
+	return newCreateVolumeResponse(disk, responseCtx), nil
 }
 
 func validateCreateVolumeRequest(req *csi.CreateVolumeRequest) error {
@@ -288,22 +365,18 @@ func validateCreateVolumeRequest(req *csi.CreateVolumeRequest) error {
 	}
 
 	if !isValidVolumeCapabilities(volCaps) {
-		modes := util.GetAccessModes(volCaps)
-		stringModes := strings.Join(*modes, ", ")
-		errString := "Volume capabilities " + stringModes + " not supported. Only AccessModes[ReadWriteOnce] supported."
-		return status.Error(codes.InvalidArgument, errString)
+		return status.Error(codes.InvalidArgument, "Volume capabilities not supported")
 	}
 	return nil
 }
 
 func (d *controllerService) DeleteVolume(ctx context.Context, req *csi.DeleteVolumeRequest) (*csi.DeleteVolumeResponse, error) {
-	klog.V(4).Infof("DeleteVolume: called with args: %+v", *req)
+	klog.V(4).InfoS("DeleteVolume: called", "args", *req)
 	if err := validateDeleteVolumeRequest(req); err != nil {
 		return nil, err
 	}
 
 	volumeID := req.GetVolumeId()
-
 	// check if a request is already in-flight
 	if ok := d.inFlight.Insert(volumeID); !ok {
 		msg := fmt.Sprintf(internal.VolumeOperationAlreadyExistsErrorMsg, volumeID)
@@ -312,8 +385,8 @@ func (d *controllerService) DeleteVolume(ctx context.Context, req *csi.DeleteVol
 	defer d.inFlight.Delete(volumeID)
 
 	if _, err := d.cloud.DeleteDisk(ctx, volumeID); err != nil {
-		if err == cloud.ErrNotFound {
-			klog.V(4).Info("DeleteVolume: volume not found, returning with success")
+		if errors.Is(err, cloud.ErrNotFound) {
+			klog.V(4).InfoS("DeleteVolume: volume not found, returning with success")
 			return &csi.DeleteVolumeResponse{}, nil
 		}
 		return nil, status.Errorf(codes.Internal, "Could not delete volume ID %q: %v", volumeID, err)
@@ -330,7 +403,7 @@ func validateDeleteVolumeRequest(req *csi.DeleteVolumeRequest) error {
 }
 
 func (d *controllerService) ControllerPublishVolume(ctx context.Context, req *csi.ControllerPublishVolumeRequest) (*csi.ControllerPublishVolumeResponse, error) {
-	klog.V(4).Infof("ControllerPublishVolume: called with args %+v", *req)
+	klog.V(4).InfoS("ControllerPublishVolume: called", "args", *req)
 	if err := validateControllerPublishVolumeRequest(req); err != nil {
 		return nil, err
 	}
@@ -338,27 +411,17 @@ func (d *controllerService) ControllerPublishVolume(ctx context.Context, req *cs
 	volumeID := req.GetVolumeId()
 	nodeID := req.GetNodeId()
 
-	if !d.cloud.IsExistInstance(ctx, nodeID) {
-		return nil, status.Errorf(codes.NotFound, "Instance %q not found", nodeID)
+	if !d.inFlight.Insert(volumeID + nodeID) {
+		return nil, status.Error(codes.Aborted, fmt.Sprintf(internal.VolumeOperationAlreadyExistsErrorMsg, volumeID))
 	}
-	disk, err := d.cloud.GetDiskByID(ctx, volumeID)
-	if err != nil {
-		if err == cloud.ErrNotFound {
-			return nil, status.Error(codes.NotFound, "Volume not found")
-		}
-		return nil, status.Errorf(codes.Internal, "Could not get volume with ID %q: %v", volumeID, err)
-	}
+	defer d.inFlight.Delete(volumeID + nodeID)
 
-	// If given volumeId already assigned to given node, will directly return current device path
+	klog.V(2).InfoS("ControllerPublishVolume: attaching", "volumeID", volumeID, "nodeID", nodeID)
 	devicePath, err := d.cloud.AttachDisk(ctx, volumeID, nodeID)
 	if err != nil {
-		if err == cloud.ErrVolumeInUse {
-			return nil, status.Error(codes.FailedPrecondition, strings.Join(disk.Attachments, ","))
-		}
-		// TODO: Check volume capability matches for ALREADY_EXISTS
 		return nil, status.Errorf(codes.Internal, "Could not attach volume %q to node %q: %v", volumeID, nodeID, err)
 	}
-	klog.V(5).Infof("[Debug] ControllerPublishVolume: volume %s attached to node %s through device %s", volumeID, nodeID, devicePath)
+	klog.InfoS("ControllerPublishVolume: attached", "volumeID", volumeID, "nodeID", nodeID, "devicePath", devicePath)
 
 	pvInfo := map[string]string{DevicePathKey: devicePath}
 	return &csi.ControllerPublishVolumeResponse{PublishContext: pvInfo}, nil
@@ -378,18 +441,15 @@ func validateControllerPublishVolumeRequest(req *csi.ControllerPublishVolumeRequ
 		return status.Error(codes.InvalidArgument, "Volume capability not provided")
 	}
 
-	caps := []*csi.VolumeCapability{volCap}
-	if !isValidVolumeCapabilities(caps) {
-		modes := util.GetAccessModes(caps)
-		stringModes := strings.Join(*modes, ", ")
-		errString := "Volume capabilities " + stringModes + " not supported. Only AccessModes[ReadWriteOnce] supported."
-		return status.Error(codes.InvalidArgument, errString)
+	if !isValidCapability(volCap) {
+		return status.Error(codes.InvalidArgument, "Volume capability not supported")
 	}
 	return nil
 }
 
 func (d *controllerService) ControllerUnpublishVolume(ctx context.Context, req *csi.ControllerUnpublishVolumeRequest) (*csi.ControllerUnpublishVolumeResponse, error) {
-	klog.V(4).Infof("ControllerUnpublishVolume: called with args %+v", *req)
+	klog.V(4).InfoS("ControllerUnpublishVolume: called", "args", *req)
+
 	if err := validateControllerUnpublishVolumeRequest(req); err != nil {
 		return nil, err
 	}
@@ -397,13 +457,20 @@ func (d *controllerService) ControllerUnpublishVolume(ctx context.Context, req *
 	volumeID := req.GetVolumeId()
 	nodeID := req.GetNodeId()
 
+	if !d.inFlight.Insert(volumeID + nodeID) {
+		return nil, status.Error(codes.Aborted, fmt.Sprintf(internal.VolumeOperationAlreadyExistsErrorMsg, volumeID))
+	}
+	defer d.inFlight.Delete(volumeID + nodeID)
+
+	klog.V(2).InfoS("ControllerUnpublishVolume: detaching", "volumeID", volumeID, "nodeID", nodeID)
 	if err := d.cloud.DetachDisk(ctx, volumeID, nodeID); err != nil {
-		if err == cloud.ErrNotFound {
+		if errors.Is(err, cloud.ErrNotFound) {
+			klog.InfoS("ControllerUnpublishVolume: attachment not found", "volumeID", volumeID, "nodeID", nodeID)
 			return &csi.ControllerUnpublishVolumeResponse{}, nil
 		}
 		return nil, status.Errorf(codes.Internal, "Could not detach volume %q from node %q: %v", volumeID, nodeID, err)
 	}
-	klog.V(5).Infof("[Debug] ControllerUnpublishVolume: volume %s detached from node %s", volumeID, nodeID)
+	klog.InfoS("ControllerUnpublishVolume: detached", "volumeID", volumeID, "nodeID", nodeID)
 
 	return &csi.ControllerUnpublishVolumeResponse{}, nil
 }
@@ -421,7 +488,7 @@ func validateControllerUnpublishVolumeRequest(req *csi.ControllerUnpublishVolume
 }
 
 func (d *controllerService) ControllerGetCapabilities(ctx context.Context, req *csi.ControllerGetCapabilitiesRequest) (*csi.ControllerGetCapabilitiesResponse, error) {
-	klog.V(4).Infof("ControllerGetCapabilities: called with args %+v", *req)
+	klog.V(4).InfoS("ControllerGetCapabilities: called", "args", *req)
 	var caps []*csi.ControllerServiceCapability
 	for _, cap := range controllerCaps {
 		c := &csi.ControllerServiceCapability{
@@ -437,17 +504,17 @@ func (d *controllerService) ControllerGetCapabilities(ctx context.Context, req *
 }
 
 func (d *controllerService) GetCapacity(ctx context.Context, req *csi.GetCapacityRequest) (*csi.GetCapacityResponse, error) {
-	klog.V(4).Infof("GetCapacity: called with args %+v", *req)
+	klog.V(4).InfoS("GetCapacity: called", "args", *req)
 	return nil, status.Error(codes.Unimplemented, "")
 }
 
 func (d *controllerService) ListVolumes(ctx context.Context, req *csi.ListVolumesRequest) (*csi.ListVolumesResponse, error) {
-	klog.V(4).Infof("ListVolumes: called with args %+v", *req)
+	klog.V(4).InfoS("ListVolumes: called", "args", *req)
 	return nil, status.Error(codes.Unimplemented, "")
 }
 
 func (d *controllerService) ValidateVolumeCapabilities(ctx context.Context, req *csi.ValidateVolumeCapabilitiesRequest) (*csi.ValidateVolumeCapabilitiesResponse, error) {
-	klog.V(4).Infof("ValidateVolumeCapabilities: called with args %+v", *req)
+	klog.V(4).InfoS("ValidateVolumeCapabilities: called", "args", *req)
 	volumeID := req.GetVolumeId()
 	if len(volumeID) == 0 {
 		return nil, status.Error(codes.InvalidArgument, "Volume ID not provided")
@@ -459,7 +526,7 @@ func (d *controllerService) ValidateVolumeCapabilities(ctx context.Context, req 
 	}
 
 	if _, err := d.cloud.GetDiskByID(ctx, volumeID); err != nil {
-		if err == cloud.ErrNotFound {
+		if errors.Is(err, cloud.ErrNotFound) {
 			return nil, status.Error(codes.NotFound, "Volume not found")
 		}
 		return nil, status.Errorf(codes.Internal, "Could not get volume with ID %q: %v", volumeID, err)
@@ -475,7 +542,7 @@ func (d *controllerService) ValidateVolumeCapabilities(ctx context.Context, req 
 }
 
 func (d *controllerService) ControllerExpandVolume(ctx context.Context, req *csi.ControllerExpandVolumeRequest) (*csi.ControllerExpandVolumeResponse, error) {
-	klog.V(4).Infof("ControllerExpandVolume: called with args %+v", *req)
+	klog.V(4).InfoS("ControllerExpandVolume: called", "args", *req)
 	volumeID := req.GetVolumeId()
 	if len(volumeID) == 0 {
 		return nil, status.Error(codes.InvalidArgument, "Volume ID not provided")
@@ -492,9 +559,26 @@ func (d *controllerService) ControllerExpandVolume(ctx context.Context, req *csi
 		return nil, status.Error(codes.InvalidArgument, "After round-up, volume size exceeds the limit specified")
 	}
 
-	actualSizeGiB, err := d.cloud.ResizeDisk(ctx, volumeID, newSize)
-	if err != nil {
-		return nil, status.Errorf(codes.Internal, "Could not resize volume %q: %v", volumeID, err)
+	responseChan := make(chan modifyVolumeResponse)
+	modifyVolumeRequest := modifyVolumeRequest{
+		newSize:      newSize,
+		responseChan: responseChan,
+	}
+
+	// Intentionally not pass in context as we deal with context locally in this method
+	d.addModifyVolumeRequest(volumeID, &modifyVolumeRequest) //nolint:contextcheck
+
+	var actualSizeGiB int64
+
+	select {
+	case response := <-responseChan:
+		if response.err != nil {
+			return nil, status.Errorf(codes.Internal, "Could not resize volume %q: %v", volumeID, response.err)
+		} else {
+			actualSizeGiB = response.volumeSize
+		}
+	case <-ctx.Done():
+		return nil, status.Errorf(codes.Internal, "Could not resize volume %q: context cancelled", volumeID)
 	}
 
 	nodeExpansionRequired := true
@@ -511,27 +595,44 @@ func (d *controllerService) ControllerExpandVolume(ctx context.Context, req *csi
 }
 
 func (d *controllerService) ControllerGetVolume(ctx context.Context, req *csi.ControllerGetVolumeRequest) (*csi.ControllerGetVolumeResponse, error) {
-	klog.V(4).Infof("ControllerGetVolume: called with args %+v", *req)
+	klog.V(4).InfoS("ControllerGetVolume: called", "args", *req)
 	return nil, status.Error(codes.Unimplemented, "")
 }
 
-func isValidVolumeCapabilities(volCaps []*csi.VolumeCapability) bool {
-	hasSupport := func(cap *csi.VolumeCapability) bool {
-		for _, c := range volumeCaps {
-			if c.GetMode() == cap.AccessMode.GetMode() {
-				return true
-			}
+func isValidVolumeCapabilities(v []*csi.VolumeCapability) bool {
+	for _, c := range v {
+		if !isValidCapability(c) {
+			return false
 		}
+	}
+	return true
+}
+
+func isValidCapability(c *csi.VolumeCapability) bool {
+	accessMode := c.GetAccessMode().GetMode()
+
+	//nolint:exhaustive
+	switch accessMode {
+	case SingleNodeWriter:
+		return true
+
+	case MultiNodeMultiWriter:
+		if isBlock(c) {
+			return true
+		} else {
+			klog.InfoS("isValidCapability: access mode is only supported for block devices", "accessMode", accessMode)
+			return false
+		}
+
+	default:
+		klog.InfoS("isValidCapability: access mode is not supported", "accessMode", accessMode)
 		return false
 	}
+}
 
-	foundAll := true
-	for _, c := range volCaps {
-		if !hasSupport(c) {
-			foundAll = false
-		}
-	}
-	return foundAll
+func isBlock(cap *csi.VolumeCapability) bool {
+	_, isBlock := cap.GetAccessType().(*csi.VolumeCapability_Block)
+	return isBlock
 }
 
 func isValidVolumeContext(volContext map[string]string) bool {
@@ -540,11 +641,11 @@ func isValidVolumeContext(volContext map[string]string) bool {
 	if partition, ok := volContext[VolumeAttributePartition]; ok {
 		partitionInt, err := strconv.ParseInt(partition, 10, 64)
 		if err != nil {
-			klog.Errorf("failed to parse partition %s as int", partition)
+			klog.ErrorS(err, "failed to parse partition as int", "partition", partition)
 			return false
 		}
 		if partitionInt < 0 {
-			klog.Errorf("invalid partition config, partition = %s", partition)
+			klog.ErrorS(err, "invalid partition config", "partition", partition)
 			return false
 		}
 	}
@@ -552,7 +653,7 @@ func isValidVolumeContext(volContext map[string]string) bool {
 }
 
 func (d *controllerService) CreateSnapshot(ctx context.Context, req *csi.CreateSnapshotRequest) (*csi.CreateSnapshotResponse, error) {
-	klog.V(4).Infof("CreateSnapshot: called with args %+v", req)
+	klog.V(4).InfoS("CreateSnapshot: called", "args", req)
 	if err := validateCreateSnapshotRequest(req); err != nil {
 		return nil, err
 	}
@@ -568,15 +669,15 @@ func (d *controllerService) CreateSnapshot(ctx context.Context, req *csi.CreateS
 	defer d.inFlight.Delete(snapshotName)
 
 	snapshot, err := d.cloud.GetSnapshotByName(ctx, snapshotName)
-	if err != nil && err != cloud.ErrNotFound {
-		klog.Errorf("Error looking for the snapshot %s: %v", snapshotName, err)
+	if err != nil && !errors.Is(err, cloud.ErrNotFound) {
+		klog.ErrorS(err, "Error looking for the snapshot", "snapshotName", snapshotName)
 		return nil, err
 	}
 	if snapshot != nil {
 		if snapshot.SourceVolumeID != volumeID {
 			return nil, status.Errorf(codes.AlreadyExists, "Snapshot %s already exists for different volume (%s)", snapshotName, snapshot.SourceVolumeID)
 		}
-		klog.V(4).Infof("Snapshot %s of volume %s already exists; nothing to do", snapshotName, volumeID)
+		klog.V(4).InfoS("Snapshot of volume already exists; nothing to do", "snapshotName", snapshotName, "volumeId", volumeID)
 		return newCreateSnapshotResponse(snapshot)
 	}
 
@@ -586,25 +687,35 @@ func (d *controllerService) CreateSnapshot(ctx context.Context, req *csi.CreateS
 	}
 
 	var vscTags []string
+	var fsrAvailabilityZones []string
+	vsProps := new(template.VolumeSnapshotProps)
 	for key, value := range req.GetParameters() {
-		if strings.HasPrefix(key, TagKeyPrefix) {
-			vscTags = append(vscTags, value)
-		} else {
-			return nil, status.Errorf(codes.InvalidArgument, "Invalid parameter key %s for CreateSnapshot", key)
+		switch strings.ToLower(key) {
+		case VolumeSnapshotNameKey:
+			vsProps.VolumeSnapshotName = value
+		case VolumeSnapshotNamespaceKey:
+			vsProps.VolumeSnapshotNamespace = value
+		case VolumeSnapshotContentNameKey:
+			vsProps.VolumeSnapshotContentName = value
+		case FastSnapshotRestoreAvailabilityZones:
+			f := strings.ReplaceAll(value, " ", "")
+			fsrAvailabilityZones = strings.Split(f, ",")
+		default:
+			if strings.HasPrefix(key, TagKeyPrefix) {
+				vscTags = append(vscTags, value)
+			} else {
+				return nil, status.Errorf(codes.InvalidArgument, "Invalid parameter key %s for CreateSnapshot", key)
+			}
 		}
 	}
 
-	addTags, err := template.Evaluate(vscTags, nil, d.driverOptions.warnOnInvalidTag)
+	addTags, err := template.Evaluate(vscTags, vsProps, d.driverOptions.warnOnInvalidTag)
 	if err != nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Error interpolating the tag value: %v", err)
 	}
 
-	if err := validateExtraTags(addTags, d.driverOptions.warnOnInvalidTag); err != nil {
+	if err = validateExtraTags(addTags, d.driverOptions.warnOnInvalidTag); err != nil {
 		return nil, status.Errorf(codes.InvalidArgument, "Invalid tag value: %v", err)
-	}
-
-	for k, v := range addTags {
-		snapshotTags[k] = v
 	}
 
 	if d.driverOptions.kubernetesClusterID != "" {
@@ -615,14 +726,43 @@ func (d *controllerService) CreateSnapshot(ctx context.Context, req *csi.CreateS
 	for k, v := range d.driverOptions.extraTags {
 		snapshotTags[k] = v
 	}
+
+	for k, v := range addTags {
+		snapshotTags[k] = v
+	}
+
 	opts := &cloud.SnapshotOptions{
 		Tags: snapshotTags,
 	}
 
-	snapshot, err = d.cloud.CreateSnapshot(ctx, volumeID, opts)
+	// Check if the availability zone is supported for fast snapshot restore
+	if len(fsrAvailabilityZones) > 0 {
+		zones, error := d.cloud.AvailabilityZones(ctx)
+		if error != nil {
+			klog.ErrorS(error, "failed to get availability zones")
+		} else {
+			klog.V(4).InfoS("Availability Zones", "zone", zones)
+			for _, az := range fsrAvailabilityZones {
+				if _, ok := zones[az]; !ok {
+					return nil, status.Errorf(codes.InvalidArgument, "Availability zone %s is not supported for fast snapshot restore", az)
+				}
+			}
+		}
+	}
 
+	snapshot, err = d.cloud.CreateSnapshot(ctx, volumeID, opts)
 	if err != nil {
 		return nil, status.Errorf(codes.Internal, "Could not create snapshot %q: %v", snapshotName, err)
+	}
+
+	if len(fsrAvailabilityZones) > 0 {
+		_, err := d.cloud.EnableFastSnapshotRestores(ctx, fsrAvailabilityZones, snapshot.SnapshotID)
+		if err != nil {
+			if _, err = d.cloud.DeleteSnapshot(ctx, snapshot.SnapshotID); err != nil {
+				return nil, status.Errorf(codes.Internal, "Could not delete snapshot ID %q: %v", snapshotName, err)
+			}
+			return nil, status.Errorf(codes.Internal, "Failed to create Fast Snapshot Restores for snapshot ID %q: %v", snapshotName, err)
+		}
 	}
 	return newCreateSnapshotResponse(snapshot)
 }
@@ -639,7 +779,7 @@ func validateCreateSnapshotRequest(req *csi.CreateSnapshotRequest) error {
 }
 
 func (d *controllerService) DeleteSnapshot(ctx context.Context, req *csi.DeleteSnapshotRequest) (*csi.DeleteSnapshotResponse, error) {
-	klog.V(4).Infof("DeleteSnapshot: called with args %+v", req)
+	klog.V(4).InfoS("DeleteSnapshot: called", "args", req)
 	if err := validateDeleteSnapshotRequest(req); err != nil {
 		return nil, err
 	}
@@ -654,8 +794,8 @@ func (d *controllerService) DeleteSnapshot(ctx context.Context, req *csi.DeleteS
 	defer d.inFlight.Delete(snapshotID)
 
 	if _, err := d.cloud.DeleteSnapshot(ctx, snapshotID); err != nil {
-		if err == cloud.ErrNotFound {
-			klog.V(4).Info("DeleteSnapshot: snapshot not found, returning with success")
+		if errors.Is(err, cloud.ErrNotFound) {
+			klog.V(4).InfoS("DeleteSnapshot: snapshot not found, returning with success")
 			return &csi.DeleteSnapshotResponse{}, nil
 		}
 		return nil, status.Errorf(codes.Internal, "Could not delete snapshot ID %q: %v", snapshotID, err)
@@ -672,27 +812,24 @@ func validateDeleteSnapshotRequest(req *csi.DeleteSnapshotRequest) error {
 }
 
 func (d *controllerService) ListSnapshots(ctx context.Context, req *csi.ListSnapshotsRequest) (*csi.ListSnapshotsResponse, error) {
-	klog.V(4).Infof("ListSnapshots: called with args %+v", req)
+	klog.V(4).InfoS("ListSnapshots: called", "args", req)
 	var snapshots []*cloud.Snapshot
 
 	snapshotID := req.GetSnapshotId()
 	if len(snapshotID) != 0 {
 		snapshot, err := d.cloud.GetSnapshotByID(ctx, snapshotID)
 		if err != nil {
-			if err == cloud.ErrNotFound {
-				klog.V(4).Info("ListSnapshots: snapshot not found, returning with success")
+			if errors.Is(err, cloud.ErrNotFound) {
+				klog.V(4).InfoS("ListSnapshots: snapshot not found, returning with success")
 				return &csi.ListSnapshotsResponse{}, nil
 			}
 			return nil, status.Errorf(codes.Internal, "Could not get snapshot ID %q: %v", snapshotID, err)
 		}
 		snapshots = append(snapshots, snapshot)
-		if response, err := newListSnapshotsResponse(&cloud.ListSnapshotsResponse{
+		response := newListSnapshotsResponse(&cloud.ListSnapshotsResponse{
 			Snapshots: snapshots,
-		}); err != nil {
-			return nil, status.Errorf(codes.Internal, "Could not build ListSnapshotsResponse: %v", err)
-		} else {
-			return response, nil
-		}
+		})
+		return response, nil
 	}
 
 	volumeID := req.GetSourceVolumeId()
@@ -701,20 +838,17 @@ func (d *controllerService) ListSnapshots(ctx context.Context, req *csi.ListSnap
 
 	cloudSnapshots, err := d.cloud.ListSnapshots(ctx, volumeID, maxEntries, nextToken)
 	if err != nil {
-		if err == cloud.ErrNotFound {
-			klog.V(4).Info("ListSnapshots: snapshot not found, returning with success")
+		if errors.Is(err, cloud.ErrNotFound) {
+			klog.V(4).InfoS("ListSnapshots: snapshot not found, returning with success")
 			return &csi.ListSnapshotsResponse{}, nil
 		}
-		if err == cloud.ErrInvalidMaxResults {
+		if errors.Is(err, cloud.ErrInvalidMaxResults) {
 			return nil, status.Errorf(codes.InvalidArgument, "Error mapping MaxEntries to AWS MaxResults: %v", err)
 		}
 		return nil, status.Errorf(codes.Internal, "Could not list snapshots: %v", err)
 	}
 
-	response, err := newListSnapshotsResponse(cloudSnapshots)
-	if err != nil {
-		return nil, status.Errorf(codes.Internal, "Could not build ListSnapshotsResponse: %v", err)
-	}
+	response := newListSnapshotsResponse(cloudSnapshots)
 	return response, nil
 }
 
@@ -768,7 +902,7 @@ func getOutpostArn(requirement *csi.TopologyRequirement) string {
 	return ""
 }
 
-func newCreateVolumeResponse(disk *cloud.Disk) *csi.CreateVolumeResponse {
+func newCreateVolumeResponse(disk *cloud.Disk, ctx map[string]string) *csi.CreateVolumeResponse {
 	var src *csi.VolumeContentSource
 	if disk.SnapshotID != "" {
 		src = &csi.VolumeContentSource{
@@ -795,7 +929,7 @@ func newCreateVolumeResponse(disk *cloud.Disk) *csi.CreateVolumeResponse {
 		Volume: &csi.Volume{
 			VolumeId:      disk.VolumeID,
 			CapacityBytes: util.GiBToBytes(disk.CapacityGiB),
-			VolumeContext: map[string]string{},
+			VolumeContext: ctx,
 			AccessibleTopology: []*csi.Topology{
 				{
 					Segments: segments,
@@ -807,10 +941,8 @@ func newCreateVolumeResponse(disk *cloud.Disk) *csi.CreateVolumeResponse {
 }
 
 func newCreateSnapshotResponse(snapshot *cloud.Snapshot) (*csi.CreateSnapshotResponse, error) {
-	ts, err := ptypes.TimestampProto(snapshot.CreationTime)
-	if err != nil {
-		return nil, err
-	}
+	ts := timestamppb.New(snapshot.CreationTime)
+
 	return &csi.CreateSnapshotResponse{
 		Snapshot: &csi.Snapshot{
 			SnapshotId:     snapshot.SnapshotID,
@@ -822,27 +954,22 @@ func newCreateSnapshotResponse(snapshot *cloud.Snapshot) (*csi.CreateSnapshotRes
 	}, nil
 }
 
-func newListSnapshotsResponse(cloudResponse *cloud.ListSnapshotsResponse) (*csi.ListSnapshotsResponse, error) {
+func newListSnapshotsResponse(cloudResponse *cloud.ListSnapshotsResponse) *csi.ListSnapshotsResponse {
 
 	var entries []*csi.ListSnapshotsResponse_Entry
 	for _, snapshot := range cloudResponse.Snapshots {
-		snapshotResponseEntry, err := newListSnapshotsResponseEntry(snapshot)
-		if err != nil {
-			return nil, err
-		}
+		snapshotResponseEntry := newListSnapshotsResponseEntry(snapshot)
 		entries = append(entries, snapshotResponseEntry)
 	}
 	return &csi.ListSnapshotsResponse{
 		Entries:   entries,
 		NextToken: cloudResponse.NextToken,
-	}, nil
+	}
 }
 
-func newListSnapshotsResponseEntry(snapshot *cloud.Snapshot) (*csi.ListSnapshotsResponse_Entry, error) {
-	ts, err := ptypes.TimestampProto(snapshot.CreationTime)
-	if err != nil {
-		return nil, err
-	}
+func newListSnapshotsResponseEntry(snapshot *cloud.Snapshot) *csi.ListSnapshotsResponse_Entry {
+	ts := timestamppb.New(snapshot.CreationTime)
+
 	return &csi.ListSnapshotsResponse_Entry{
 		Snapshot: &csi.Snapshot{
 			SnapshotId:     snapshot.SnapshotID,
@@ -851,7 +978,7 @@ func newListSnapshotsResponseEntry(snapshot *cloud.Snapshot) (*csi.ListSnapshots
 			CreationTime:   ts,
 			ReadyToUse:     snapshot.ReadyToUse,
 		},
-	}, nil
+	}
 }
 
 func getVolSizeBytes(req *csi.CreateVolumeRequest) (int64, error) {
@@ -892,4 +1019,25 @@ func BuildOutpostArn(segments map[string]string) string {
 		segments[AwsAccountIDKey],
 		segments[AwsOutpostIDKey],
 	)
+}
+
+func validateFormattingOption(volumeCapabilities []*csi.VolumeCapability, paramName string, fsConfigs map[string]fileSystemConfig) error {
+	for _, volCap := range volumeCapabilities {
+		switch volCap.GetAccessType().(type) {
+		case *csi.VolumeCapability_Block:
+			return status.Error(codes.InvalidArgument, fmt.Sprintf("Cannot use %s with block volume", paramName))
+		}
+
+		mountVolume := volCap.GetMount()
+		if mountVolume == nil {
+			return status.Error(codes.InvalidArgument, "CreateVolume: mount is nil within volume capability")
+		}
+
+		fsType := mountVolume.GetFsType()
+		if supported := fsConfigs[fsType].isParameterSupported(paramName); !supported {
+			return status.Errorf(codes.InvalidArgument, "Cannot use %s with fstype %s", paramName, fsType)
+		}
+	}
+
+	return nil
 }
