@@ -21,8 +21,10 @@ import (
 	"fmt"
 	"net"
 
+	"github.com/awslabs/volume-modifier-for-k8s/pkg/rpc"
 	csi "github.com/container-storage-interface/spec/lib/go/csi"
 	"github.com/kubernetes-sigs/aws-ebs-csi-driver/pkg/util"
+	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
 	"google.golang.org/grpc"
 	"k8s.io/klog/v2"
 )
@@ -66,11 +68,14 @@ type DriverOptions struct {
 	volumeAttachLimit   int64
 	kubernetesClusterID string
 	awsSdkDebugLog      bool
+	batching            bool
 	warnOnInvalidTag    bool
+	userAgentExtra      string
+	otelTracing         bool
 }
 
 func NewDriver(options ...func(*DriverOptions)) (*Driver, error) {
-	klog.V(4).Infof("Driver: %v Version: %v", DriverName, driverVersion)
+	klog.InfoS("Driver Information", "Driver", DriverName, "Version", driverVersion)
 
 	driverOptions := DriverOptions{
 		endpoint: DefaultCSIEndpoint,
@@ -81,7 +86,7 @@ func NewDriver(options ...func(*DriverOptions)) (*Driver, error) {
 	}
 
 	if err := ValidateDriverOptions(&driverOptions); err != nil {
-		return nil, fmt.Errorf("Invalid driver options: %v", err)
+		return nil, fmt.Errorf("Invalid driver options: %w", err)
 	}
 
 	driver := Driver{
@@ -117,12 +122,18 @@ func (d *Driver) Run() error {
 	logErr := func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
 		resp, err := handler(ctx, req)
 		if err != nil {
-			klog.Errorf("GRPC error: %v", err)
+			klog.ErrorS(err, "GRPC error")
 		}
 		return resp, err
 	}
+
+	grpcInterceptor := grpc.UnaryInterceptor(logErr)
+	if d.options.otelTracing {
+		grpcInterceptor = grpc.ChainUnaryInterceptor(logErr, otelgrpc.UnaryServerInterceptor())
+	}
+
 	opts := []grpc.ServerOption{
-		grpc.UnaryInterceptor(logErr),
+		grpcInterceptor,
 	}
 	d.srv = grpc.NewServer(opts...)
 
@@ -131,16 +142,18 @@ func (d *Driver) Run() error {
 	switch d.options.mode {
 	case ControllerMode:
 		csi.RegisterControllerServer(d.srv, d)
+		rpc.RegisterModifyServer(d.srv, d)
 	case NodeMode:
 		csi.RegisterNodeServer(d.srv, d)
 	case AllMode:
 		csi.RegisterControllerServer(d.srv, d)
 		csi.RegisterNodeServer(d.srv, d)
+		rpc.RegisterModifyServer(d.srv, d)
 	default:
 		return fmt.Errorf("unknown mode: %s", d.options.mode)
 	}
 
-	klog.V(4).Infof("Listening for connections on address: %#v", listener.Addr())
+	klog.V(4).InfoS("Listening for connections", "address", listener.Addr())
 	return d.srv.Serve(listener)
 }
 
@@ -163,7 +176,7 @@ func WithExtraTags(extraTags map[string]string) func(*DriverOptions) {
 func WithExtraVolumeTags(extraVolumeTags map[string]string) func(*DriverOptions) {
 	return func(o *DriverOptions) {
 		if o.extraTags == nil && extraVolumeTags != nil {
-			klog.Warning("DEPRECATION WARNING: --extra-volume-tags is deprecated, please use --extra-tags instead")
+			klog.InfoS("DEPRECATION WARNING: --extra-volume-tags is deprecated, please use --extra-tags instead")
 			o.extraTags = extraVolumeTags
 		}
 	}
@@ -178,6 +191,12 @@ func WithMode(mode Mode) func(*DriverOptions) {
 func WithVolumeAttachLimit(volumeAttachLimit int64) func(*DriverOptions) {
 	return func(o *DriverOptions) {
 		o.volumeAttachLimit = volumeAttachLimit
+	}
+}
+
+func WithBatching(enableBatching bool) func(*DriverOptions) {
+	return func(o *DriverOptions) {
+		o.batching = enableBatching
 	}
 }
 
@@ -196,5 +215,17 @@ func WithAwsSdkDebugLog(enableSdkDebugLog bool) func(*DriverOptions) {
 func WithWarnOnInvalidTag(warnOnInvalidTag bool) func(*DriverOptions) {
 	return func(o *DriverOptions) {
 		o.warnOnInvalidTag = warnOnInvalidTag
+	}
+}
+
+func WithUserAgentExtra(userAgentExtra string) func(*DriverOptions) {
+	return func(o *DriverOptions) {
+		o.userAgentExtra = userAgentExtra
+	}
+}
+
+func WithOtelTracing(enableOtelTracing bool) func(*DriverOptions) {
+	return func(o *DriverOptions) {
+		o.otelTracing = enableOtelTracing
 	}
 }
