@@ -24,8 +24,6 @@ import (
 	mountutils "k8s.io/mount-utils"
 )
 
-type mountInterface = mountutils.Interface
-
 // Mounter is the interface implemented by NodeMounter.
 // A mix & match of functions defined in upstream libraries. (FormatAndMount
 // from struct SafeFormatAndMount, PathExists from an old edition of
@@ -34,7 +32,7 @@ type mountInterface = mountutils.Interface
 type Mounter interface {
 	mountutils.Interface
 
-	FormatAndMount(source string, target string, fstype string, options []string) error
+	FormatAndMountSensitiveWithFormatOptions(source string, target string, fstype string, options []string, sensitiveOptions []string, formatOptions []string) error
 	IsCorruptedMnt(err error) bool
 	GetDeviceNameFromMount(mountPath string) (string, int, error)
 	MakeFile(path string) error
