@@ -57,29 +57,29 @@ func (d *nodeService) findDevicePath(devicePath, volumeID, partition string) (st
 	// path below must happen and must rely on volume ID
 	exists, err := d.mounter.PathExists(devicePath)
 	if err != nil {
-		return "", fmt.Errorf("failed to check if path %q exists: %v", devicePath, err)
+		return "", fmt.Errorf("failed to check if path %q exists: %w", devicePath, err)
 	}
 
 	if exists {
-		stat, err := d.deviceIdentifier.Lstat(devicePath)
-		if err != nil {
-			return "", fmt.Errorf("failed to lstat %q: %v", devicePath, err)
+		stat, lstatErr := d.deviceIdentifier.Lstat(devicePath)
+		if lstatErr != nil {
+			return "", fmt.Errorf("failed to lstat %q: %w", devicePath, err)
 		}
 
 		if stat.Mode()&os.ModeSymlink == os.ModeSymlink {
 			canonicalDevicePath, err = d.deviceIdentifier.EvalSymlinks(devicePath)
 			if err != nil {
-				return "", fmt.Errorf("failed to evaluate symlink %q: %v", devicePath, err)
+				return "", fmt.Errorf("failed to evaluate symlink %q: %w", devicePath, err)
 			}
 		} else {
 			canonicalDevicePath = devicePath
 		}
 
-		klog.V(5).Infof("[Debug] The canonical device path for %q was resolved to: %q", devicePath, canonicalDevicePath)
+		klog.V(5).InfoS("[Debug] The canonical device path was resolved", "devicePath", devicePath, "cacanonicalDevicePath", canonicalDevicePath)
 		return d.appendPartition(canonicalDevicePath, partition), nil
 	}
 
-	klog.V(5).Infof("[Debug] Falling back to nvme volume ID lookup for: %q", devicePath)
+	klog.V(5).InfoS("[Debug] Falling back to nvme volume ID lookup", "devicePath", devicePath)
 
 	// AWS recommends identifying devices by volume ID
 	// (https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/nvme-ebs-volumes.html),
@@ -92,16 +92,20 @@ func (d *nodeService) findDevicePath(devicePath, volumeID, partition string) (st
 	nvmeDevicePath, err := findNvmeVolume(d.deviceIdentifier, nvmeName)
 
 	if err == nil {
-		klog.V(5).Infof("[Debug] successfully resolved nvmeName=%q to %q", nvmeName, nvmeDevicePath)
+		klog.V(5).InfoS("[Debug] successfully resolved", "nvmeName", nvmeName, "nvmeDevicePath", nvmeDevicePath)
 		canonicalDevicePath = nvmeDevicePath
 		return d.appendPartition(canonicalDevicePath, partition), nil
 	} else {
-		klog.V(5).Infof("[Debug] error searching for nvme path %q: %v", nvmeName, err)
+		klog.V(5).InfoS("[Debug] error searching for nvme path", "nvmeName", nvmeName, "err", err)
 	}
 
 	if util.IsSBE(d.metadata.GetRegion()) {
-		klog.V(5).Infof("[Debug] Falling back to snow volume lookup for: %q", devicePath)
-		canonicalDevicePath = "/dev/vd" + strings.TrimPrefix(devicePath, "/dev/xvdb")
+		klog.V(5).InfoS("[Debug] Falling back to snow volume lookup", "devicePath", devicePath)
+		// Snow completely ignores the requested device path and mounts volumes starting at /dev/vda .. /dev/vdb .. etc
+		// Morph the device path to the snow form by chopping off the last letter and prefixing with /dev/vd
+		// VMs on snow devices are currently limited to 10 block devices each - if that ever exceeds 26 this will need
+		// to be adapted
+		canonicalDevicePath = "/dev/vd" + devicePath[len(devicePath)-1:]
 	}
 
 	if canonicalDevicePath == "" {
@@ -123,21 +127,21 @@ func findNvmeVolume(deviceIdentifier DeviceIdentifier, findName string) (device 
 	stat, err := deviceIdentifier.Lstat(p)
 	if err != nil {
 		if os.IsNotExist(err) {
-			klog.V(5).Infof("[Debug] nvme path %q not found", p)
+			klog.V(5).InfoS("[Debug] nvme path not found", "path", p)
 			return "", fmt.Errorf("nvme path %q not found", p)
 		}
-		return "", fmt.Errorf("error getting stat of %q: %v", p, err)
+		return "", fmt.Errorf("error getting stat of %q: %w", p, err)
 	}
 
 	if stat.Mode()&os.ModeSymlink != os.ModeSymlink {
-		klog.Warningf("nvme file %q found, but was not a symlink", p)
+		klog.InfoS("nvme file found, but was not a symlink", "path", p)
 		return "", fmt.Errorf("nvme file %q found, but was not a symlink", p)
 	}
 	// Find the target, resolving to an absolute path
 	// For example, /dev/disk/by-id/nvme-Amazon_Elastic_Block_Store_vol0fab1d5e3f72a5e23 -> ../../nvme2n1
 	resolved, err := deviceIdentifier.EvalSymlinks(p)
 	if err != nil {
-		return "", fmt.Errorf("error reading target of symlink %q: %v", p, err)
+		return "", fmt.Errorf("error reading target of symlink %q: %w", p, err)
 	}
 
 	if !strings.HasPrefix(resolved, "/dev") {
@@ -148,9 +152,9 @@ func findNvmeVolume(deviceIdentifier DeviceIdentifier, findName string) (device 
 }
 
 func (d *nodeService) preparePublishTarget(target string) error {
-	klog.V(4).Infof("NodePublishVolume: creating dir %s", target)
+	klog.V(4).InfoS("NodePublishVolume: creating dir", "target", target)
 	if err := d.mounter.MakeDir(target); err != nil {
-		return fmt.Errorf("Could not create dir %q: %v", target, err)
+		return fmt.Errorf("Could not create dir %q: %w", target, err)
 	}
 	return nil
 }
@@ -170,7 +174,7 @@ func (d *nodeService) getBlockSizeBytes(devicePath string) (int64, error) {
 	cmd := d.mounter.(*NodeMounter).Exec.Command("blockdev", "--getsize64", devicePath)
 	output, err := cmd.Output()
 	if err != nil {
-		return -1, fmt.Errorf("error when getting size of block volume at path %s: output: %s, err: %v", devicePath, string(output), err)
+		return -1, fmt.Errorf("error when getting size of block volume at path %s: output: %s, err: %w", devicePath, string(output), err)
 	}
 	strOut := strings.TrimSpace(string(output))
 	gotSizeBytes, err := strconv.ParseInt(strOut, 10, 64)
