@@ -189,10 +189,151 @@ func TestEvaluate(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 
-			props := &Props{
+			props := &PVProps{
 				PVCName:      tc.pvcName,
 				PVCNamespace: tc.pvcNamespace,
 				PVName:       tc.pvName,
+			}
+
+			tags, err := Evaluate(tc.input, props, tc.warnOnly)
+
+			if tc.expectErr {
+				if err == nil {
+					t.Fatalf("expected an error; got nil")
+				}
+			} else {
+				if err != nil {
+					t.Fatalf("err is not nil; err = %v", err)
+				}
+				if diff := cmp.Diff(tc.expectedTags, tags); diff != "" {
+					t.Fatalf("tags are different; diff = %v", diff)
+				}
+			}
+		})
+	}
+}
+
+func TestEvaluateVolumeSnapshotTemplate(t *testing.T) {
+	testCases := []struct {
+		name                      string
+		input                     []string
+		volumeSnapshotName        string
+		volumeSnapshotNamespace   string
+		volumeSnapshotContentName string
+		warnOnly                  bool
+		expectErr                 bool
+		expectedTags              map[string]string
+	}{
+		{
+			name: "simple substitution",
+			input: []string{
+				"key1={{ .VolumeSnapshotName }}",
+				"key2={{ .VolumeSnapshotNamespace }}",
+				"key3={{ .VolumeSnapshotContentName }}",
+			},
+			volumeSnapshotName:        "ebs-vs",
+			volumeSnapshotNamespace:   "default",
+			volumeSnapshotContentName: "ebs-vs-content-012345",
+			expectedTags: map[string]string{
+				"key1": "ebs-vs",
+				"key2": "default",
+				"key3": "ebs-vs-content-012345",
+			},
+		},
+		{
+			name: "template parsing error",
+			input: []string{
+				"key1={{ .VolumeSnapshotName }",
+			},
+			expectErr: true,
+		},
+		{
+			name: "template parsing error warn only",
+			input: []string{
+				"key1={{ .VolumeSnapshotName }",
+				"key2={{ .VolumeSnapshotNamespace }}",
+			},
+			volumeSnapshotName:      "ebs-vs",
+			volumeSnapshotNamespace: "default",
+			warnOnly:                true,
+			expectedTags: map[string]string{
+				"key2": "default",
+			},
+		},
+		{
+			name: "test unsupported func - returns error",
+			input: []string{
+				`backup={{ .VolumeSnapshotNamespace | html }}`,
+			},
+			volumeSnapshotNamespace: "ns-prod",
+			expectErr:               true,
+		},
+		{
+			name: "test function - contains",
+			input: []string{
+				`backup={{ .VolumeSnapshotNamespace | contains "prod" }}`,
+			},
+			volumeSnapshotNamespace: "ns-prod",
+			expectedTags: map[string]string{
+				"backup": "true",
+			},
+		},
+		{
+			name: "test function - toUpper",
+			input: []string{
+				`backup={{ .VolumeSnapshotNamespace | toUpper }}`,
+			},
+			volumeSnapshotNamespace: "ns-prod",
+			expectedTags: map[string]string{
+				"backup": "NS-PROD",
+			},
+		},
+		{
+			name: "test function - toLower",
+			input: []string{
+				`backup={{ .VolumeSnapshotNamespace | toLower }}`,
+			},
+			volumeSnapshotNamespace: "ns-PROD",
+			expectedTags: map[string]string{
+				"backup": "ns-prod",
+			},
+		},
+		{
+			name: "test function - field",
+			input: []string{
+				`backup={{ .VolumeSnapshotNamespace | field "-" 1 }}`,
+			},
+			volumeSnapshotNamespace: "ns-prod-default",
+			expectedTags: map[string]string{
+				"backup": "prod",
+			},
+		},
+		{
+			name: "test function - substring",
+			input: []string{
+				`key1={{ .VolumeSnapshotNamespace | substring 0 4 }}`,
+			},
+			volumeSnapshotNamespace: "prod-12345",
+			expectedTags: map[string]string{
+				"key1": "prod",
+			},
+		},
+		{
+			name: "field returns error",
+			input: []string{
+				`key1={{ .VolumeSnapshotNamespace | field "-" 1 }}`,
+			},
+			expectErr: true,
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+
+			props := &VolumeSnapshotProps{
+				VolumeSnapshotName:        tc.volumeSnapshotName,
+				VolumeSnapshotNamespace:   tc.volumeSnapshotNamespace,
+				VolumeSnapshotContentName: tc.volumeSnapshotContentName,
 			}
 
 			tags, err := Evaluate(tc.input, props, tc.warnOnly)
